fix: load missing total coin count from file and update tests accordingly
