+++ src/main/java/com/dinosaur/dinosaurexploder/utils/DataProvider.java
@@ -1,7 +1,10 @@
 package com.dinosaur.dinosaurexploder.utils;
 
 import com.dinosaur.dinosaurexploder.model.HighScore;
+import com.dinosaur.dinosaurexploder.model.TotalCoins;
 
 public interface DataProvider {
     HighScore getHighScore();
+
+    TotalCoins getTotalCoins();
 }
