+++ src/main/java/com/dinosaur/dinosaurexploder/utils/FileDataProvider.java
@@ -1,6 +1,7 @@
 package com.dinosaur.dinosaurexploder.utils;
 
 import com.dinosaur.dinosaurexploder.model.HighScore;
+import com.dinosaur.dinosaurexploder.model.TotalCoins;
 
 import java.io.FileInputStream;
 import java.io.IOException;
@@ -14,6 +15,16 @@ public class FileDataProvider implements DataProvider {
             return (HighScore) in.readObject();
         } catch (IOException | ClassNotFoundException e) {
             return new HighScore();
+        }
+    }
+
+    @Override
+    public TotalCoins getTotalCoins() {
+        try (FileInputStream file = new FileInputStream("totalCoins.ser");
+             ObjectInputStream in = new ObjectInputStream(file)) {
+            return (TotalCoins) in.readObject();
+        } catch (IOException | ClassNotFoundException e) {
+            return new TotalCoins();
         }
     }
 }
